+++ app/controllers/carto/api/visualizations_controller.rb
@@ -94,15 +94,12 @@ module Carto
         visualizations = vqb.with_order(order, order_direction)
                             .build_paged(page, per_page).map do |v|
           VisualizationPresenter.new(v, current_viewer, self, presenter_options)
-                                .with_presenter_cache(presenter_cache).to_poro \
-            unless (params[:subscribed] == 'true' and not v.subscription.present?) or (params[:sample] == 'true' and not v.sample.present?)
+                                .with_presenter_cache(presenter_cache).to_poro
         end.compact
 
         response = { visualizations: visualizations, total_entries: vqb.count }
         response.merge!(calculate_totals(types)) if current_user && params[:load_totals].to_s != 'false'
         response.merge!(calculate_do_totals(vqb)) if params[:load_do_totals].to_s == 'true'
-        response[:total_entries] = response[:total_subscriptions] if params[:subscribed] == 'true'
-        response[:total_entries] = response[:total_samples] if params[:sample] == 'true'
 
         render_jsonp(response)
       rescue CartoDB::BoundingBoxError => e
@@ -503,12 +500,21 @@ module Carto
       end
 
       def calculate_do_totals(vqb)
-        {
-          total_subscriptions: vqb.filtered_query.includes(map: { user_table: :data_import })
-                                  .find_each.lazy.count { |v| v.subscription.present? },
-          total_samples: vqb.filtered_query.includes(map: { user_table: :data_import })
-                            .find_each.lazy.count { |v| v.sample.present? }
-        }
+        subscription_count = if params[:subscribed] == 'true'
+                               vqb.count
+                             else
+                               vqb.filtered_query.includes(map: { user_table: :data_import }).find_each.lazy
+                                  .count { |v| v.subscription.present? }
+                             end
+
+        sample_count = if params[:sample] == 'true'
+                         vqb.count
+                       else
+                         vqb.filtered_query.includes(map: { user_table: :data_import }).find_each.lazy
+                            .count { |v| v.sample.present? }
+                       end
+
+        { total_subscriptions: subscription_count, total_samples: sample_count }
       end
 
       def log_context